+++ src/core/com/lowagie/text/pdf/codec/TIFFFaxDecoder.java
@@ -833,7 +833,7 @@ public class TIFFFaxDecoder {
                     entry = nextLesserThan8Bits(7);
                     
                     // Run these through the 2DCodes table
-                    entry = (int)(twoDCodes[entry] & 0xff);
+                    entry = twoDCodes[entry] & 0xff;
                     
                     // Get the code and the number of bits used up
                     code = (entry & 0x78) >>> 3;
@@ -978,7 +978,7 @@ public class TIFFFaxDecoder {
                 // Get the next seven bits
                 entry = nextLesserThan8Bits(7);
                 // Run these through the 2DCodes table
-                entry = (int)(twoDCodes[entry] & 0xff);
+                entry = twoDCodes[entry] & 0xff;
                 
                 // Get the code and the number of bits used up
                 code = (entry & 0x78) >>> 3;
